refactor: Allow components during an edit; safeguard from missing HTTPClient. (#433)
* refactor:

- raise error if HTTPClient is not defined in helper methods to have a clearer error
- add explanation of that error and fix to faq
- began channel.purge

* thing

* Update faq.rst

* Update channel.py

* Update guild.py

* Update member.py

* Update message.py

* Update role.py

* give context.message the HTTPClient
fix removal of message contents when editing

+++ interactions/api/models/role.py
@@ -75,7 +75,8 @@ class Role(DictSerializerMixin):
         :param reason: The reason for the deletion
         :type reason: Optional[str]
         """
-
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         await self._client.delete_guild_role(
             guild_id=guild_id, role_id=int(self.id), reason=reason
         ),
@@ -109,7 +110,8 @@ class Role(DictSerializerMixin):
         :return: The modified role object
         :rtype: Role
         """
-
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         _name = self.name if not name else name
         _color = self.color if not color else color
         _hoist = self.hoist if not hoist else hoist
@@ -143,7 +145,8 @@ class Role(DictSerializerMixin):
         :return: List of guild roles with updated hierarchy
         :rtype: List[Role]
         """
-
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         res = await self._client.modify_guild_role_position(
             guild_id=guild_id, position=position, role_id=int(self.id), reason=reason
         )
